+++ src/Cpp/1-getting-started/1-3-2-LoadingMeshes-Refactored/LoadingMeshesApplication.cpp
@@ -153,7 +153,11 @@ bool LoadingMeshesApplication::Load()
         return false;
     }
 
-    _pipeline->SetViewport(0.0f, 0.0f, GetWindowWidth(), GetWindowHeight());
+    _pipeline->SetViewport(
+        0.0f,
+        0.0f,
+        static_cast<float>(GetWindowWidth()),
+        static_cast<float>(GetWindowHeight()));
 
     if (!_textureFactory->CreateShaderResourceViewFromFile(L"Assets/Textures/T_Good_Froge.dds", _textureSrv))
     {
@@ -280,15 +284,15 @@ void LoadingMeshesApplication::OnResize(
 
 void LoadingMeshesApplication::Update()
 {
-    const auto eyePosition = DirectX::XMVectorSet(0, 50, 200, 1);
-    const auto focusPoint = DirectX::XMVectorSet(0, 0, 0, 1);
-    const auto upDirection = DirectX::XMVectorSet(0, 1, 0, 0);
+    const auto eyePosition = DirectX::XMVectorSet(0.0f, 50.0f, 200.0f, 1.0f);
+    const auto focusPoint = DirectX::XMVectorSet(0.0f, 0.0f, 0.0f, 1.0f);
+    const auto upDirection = DirectX::XMVectorSet(0.0f, 1.0f, 0.0f, 0.0f);
     _viewMatrix = DirectX::XMMatrixLookAtLH(eyePosition, focusPoint, upDirection);
     _deviceContext->UpdateSubresource(_constantBuffers[PerFrame].Get(), &_viewMatrix);
 
     static float angle = 0.0f;
-    angle += 90.0f * (10.0 / 60000.0f);
-    const auto rotationAxis = DirectX::XMVectorSet(0, 1, 0, 0);
+    angle += 90.0f * (10.0f / 60000.0f);
+    const auto rotationAxis = DirectX::XMVectorSet(0.0f, 1.0f, 0.0f, 0.0f);
 
     _worldMatrix = DirectX::XMMatrixRotationAxis(rotationAxis, DirectX::XMConvertToRadians(angle));
     _deviceContext->UpdateSubresource(_constantBuffers[PerObject].Get(), &_worldMatrix);
